OV-29: * better strcuture for helpers methods
