+++ frontend/src/bundles/common/helpers/array-helper/array-helper.ts
@@ -1,0 +1,1 @@
+export { isEmptyArray } from './is-empty-array.js';
