+++ frontend/src/bundles/common/helpers/array-helper/is-empty-array.ts
@@ -1,0 +1,6 @@
+import { isNullOrUndefined } from '../helpers.js';
+
+const isEmptyArray = <T>(value: T[]): boolean =>
+    isNullOrUndefined<T[]>(value) || value.length === 0;
+
+export { isEmptyArray };
