+++ frontend/src/bundles/common/helpers/helpers.ts
@@ -1,6 +1,4 @@
-export {
-    isEmptyArray,
-    isNullOrUndefined,
-    isStringNullOrEmpty,
-} from './utilities.js';
+export { isEmptyArray } from './array-helper/array-helper.js';
+export { isNullOrUndefined } from './is-null-or-undefined/is-null-or-undefined.js';
+export { isStringNullOrEmpty } from './string-helpers/string-helper.js';
 export { configureString } from 'shared';
