+++ frontend/src/bundles/common/helpers/is-null-or-undefined/is-null-or-undefined.ts
@@ -1,0 +1,5 @@
+const isNullOrUndefined = <T>(
+    value: T | null | undefined,
+): value is null | undefined => value === undefined || value === null;
+
+export { isNullOrUndefined };
