+++ frontend/src/bundles/common/helpers/string-helpers/is-null-or-undefined.ts
@@ -1,0 +1,8 @@
+import { isNullOrUndefined } from '../helpers.js';
+
+const isStringNullOrEmpty = (
+    value: string | null | undefined,
+): value is null | undefined =>
+    isNullOrUndefined<string>(value) || value.trim().length === 0;
+
+export { isStringNullOrEmpty };
