+++ frontend/src/bundles/common/helpers/string-helpers/string-helper.ts
@@ -1,0 +1,1 @@
+export { isStringNullOrEmpty } from './is-null-or-undefined.js';
